rename package autodetect to autodiscover
