+++ optaplanner-core/src/main/java/org/optaplanner/core/api/domain/autodiscover/AutoDiscoverMemberType.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.api.domain.autodetect;
+package org.optaplanner.core.api.domain.autodiscover;
 
 import org.optaplanner.core.api.domain.solution.PlanningEntityCollectionProperty;
 import org.optaplanner.core.api.domain.solution.PlanningEntityProperty;
