+++ optaplanner-core/src/test/java/org/optaplanner/core/impl/testdata/domain/solutionproperties/autodiscover/TestdataAutoDiscoverGetterSolution.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.testdata.domain.solutionproperties.autodetect;
+package org.optaplanner.core.impl.testdata.domain.solutionproperties.autodiscover;
 
 import java.util.List;
 
-import org.optaplanner.core.api.domain.autodetect.AutoDiscoverMemberType;
-import org.optaplanner.core.api.domain.solution.PlanningScore;
+import org.optaplanner.core.api.domain.autodiscover.AutoDiscoverMemberType;
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.domain.valuerange.ValueRangeProvider;
 import org.optaplanner.core.api.score.buildin.simple.SimpleScore;
